Split validate.py into validate_core.py and validate_ui.py
* validate_core.py is a Streamlit-free module for both this App external clients

* validate_ui.py is a wrapper of validate_core.py for this App Streamlit logging
